agent/requirements.txt
netifaces
python-daemon